API/2fa/requirements.txt
fastapi==0.99.1
pydantic==1.10.7
uvicorn
supabase
pycryptodome
python-dotenv